DB/DBManager/pyproject.toml
[tool.poetry]
name = "dbmanager"
version = "0.1.0"
description = ""
authors = ["rbwo552 <rbwo552@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
sqlalchemy = "^2.0.30"
faker = "^25.4.0"
flask-sqlalchemy = "^3.1.1"
flask-mysqldb = "^2.0.0"
pyyaml = "^6.0.1"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
